refactor: promote db.rs to db/ directory (operations + tunnel)
Split the two independent db subsystems into separate files:
- db/operations.rs: Query, search, list/describe tables, delete/drop via extension CLI
- db/tunnel.rs: SSH tunnel for forwarding local ports to remote databases
- db/mod.rs: Re-exports everything to preserve existing import paths

508 lines → two focused files. Zero behavior change.
